For an object's "known properties", don't assume an artifact inherits every property from its base type.
obj_flags would override the base item's flags with the artifact's, but
obj_flags_known would |= them instead. This would lead to weird behavior
if an artifact was missing a property that its base item normally has
(e.g. the mirror shield of Gil-Galad, which lacks rLite) -- the object
wouldn't actually have that property, but the game would lie to the
player that it did.