Revert "Allow teleporting/recalling to all locations from birth (MarvinPA)."
This reverts commit e959e351463d6e77f7ed9fdfa2da4baf0b6f9443.

It's nice for quality of life but it's also a big balance change because of the dungeon entrance guardians. Being able to e.g. sneak into Dragon Cave early without having to bypass a dracolisk is a huge change.

Maybe I'll re-add this as a birth option, or come up with some system to have the guardian monsters still be relevant.